+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- changed search_project tool to use Deno-native `walk` options for match and skip
+- updated tool description to clarify use of ** vs * in file patterns
 
 
 ## [0.3.2] - 2024-11-18